requirements.txt
fastapi==0.133.1
uvicorn==0.41.0

numpy==2.4.2
scikit-learn==1.8.0
scipy==1.17.1

faiss-cpu==1.13.2
rank-bm25==0.2.2

transformers
sentence-transformers

openai==2.24.0

pymongo==4.17.0
motor==3.7.1

PyMuPDF==1.27.1

python-dotenv==1.2.1
python-multipart==0.0.22

requests==2.34.0

Authlib==1.7.0
python-jose==3.5.0
email-validator==2.3.0

jinja2==3.1.6